feat(train-cpu): add Cargo.toml, lib.rs, fix type errors (Refs #321) (#363)
* docs: MCP setup guide, trainer README, UI ring audit (Closes #332, Closes #327, Closes #253)

#332: Add docs/mcp/MCP_SETUP_GUIDE.md — Perplexity local+remote MCP setup
#327: Add crates/trios-train-cpu/README.md — migration status + Training-Flow V2 plan
#253: Add missing RING.md/AGENTS.md/TASK.md for UR-04, UR-05, UR-06 + UR-03 TASK.md

* feat(train-cpu): add Cargo.toml, lib.rs, fix f32/f64 type errors (Refs #321)

- Add Cargo.toml with workspace deps (serde, rand, anyhow)
- Add src/lib.rs exporting mup and schedule modules
- Register trios-train-cpu in workspace members
- Fix schedule.rs f32/f64 type mismatch in interpolate()
- Fix mup.rs ambiguous float literal in test

15/15 tests pass. Refs #321 consolidation plan.

+++ crates/trios-train-cpu/src/schedule.rs
@@ -61,7 +61,7 @@ impl ScheduleFreeState {
     /// y_t = (1 - beta1) * z_t + beta1 * x_t
     pub fn interpolate(&self, x: &[f32], z: &[f32], beta1: f64) -> Vec<f32> {
         x.iter().zip(z.iter())
-            .map(|(&xi, &zi)| (1.0 - beta1) * zi + beta1 * xi)
+            .map(|(&xi, &zi)| ((1.0 - beta1) * zi as f64 + beta1 * xi as f64) as f32)
             .collect()
     }
 }
